pyproject.toml
[project]
name = "rag-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.15",
    "google-genai>=1.24.0",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.0.0",
]
